Optimize latlng->census tract mapping.
This replaces the previous mapping of lat/lng coordinates to census tract with a more optimal version. This version removes the necessity for polygons to be re-computed for every address input, storing those polygons in memory after a single computation. The code for the old version remains, however it can be removed once this has been looked at. Currently improves processing on 10 addresses from ~10s to ~3s.