An example of using the census.gov geocoder
On one hand, this goes directly from address → tract without an
intervening step or any local data and is really simple.  Nice!  On the
other hand, it doesn't have great address coverage, particularly in
non-residential areas, which is indeed noted as a potential pitfall by
the Census' documentation.  Whomp.

A two-step process of trying to get any lat/long (which might be
variable precision depending on the address) and then converting that to
a census tract seems more robust to geocoding problems.  The Census is
concerned with fully-specified addresses to get block and county
information too, which we don't care about.

One pitfall is that geocoding ambiguity (i.e. we have an address but the
geocoder can only place it at Seattle's centroid) is that we get an
artificial abundance of addresses assigned to the tract in which
low-geographic resolution centroids fall.  c.f. Null Island.